wip: refactor feedbacks logic into FragmentFeedbacks
