+++ lib/Controls/ControlTypes/Button/Base.js
@@ -1,8 +1,8 @@
 import ControlBase from '../../ControlBase.js'
-import { ParseControlId, rgb, clamp, GetButtonBitmapSize } from '../../../Resources/Util.js'
-import { cloneDeep, isEqual } from 'lodash-es'
-import { nanoid } from 'nanoid'
+import { ParseControlId, rgb, GetButtonBitmapSize } from '../../../Resources/Util.js'
+import { cloneDeep } from 'lodash-es'
 import FragmentActions from '../../Fragments/FragmentActions.js'
+import FragmentFeedbacks from '../../Fragments/FragmentFeedbacks.js'
 
 /**
  * Abstract class for a editable button control.
@@ -60,16 +60,16 @@ export default class ButtonControlBase extends ControlBase {
 	actions
 
 	/**
+	 * The feedbacks fragment
+	 * @access public
+	 */
+	feedbacks
+
+	/**
 	 * The current status of this button
 	 * @access protected
 	 */
 	bank_status = 'good'
-
-	/**
-	 * Cached values for the feedbacks on this control
-	 * @access private
-	 */
-	#cachedFeedbackValues = {}
 
 	/**
 	 * The base style of this button
@@ -81,12 +81,6 @@ export default class ButtonControlBase extends ControlBase {
 	 * The config of this button
 	 */
 	options = {}
-
-	/**
-	 * The feedbacks on this button
-	 * @access public
-	 */
-	feedbacks = []
 
 	/**
 	 * Whether this button has delayed actions running
@@ -108,6 +102,13 @@ export default class ButtonControlBase extends ControlBase {
 			controlId,
 			this.commitChange.bind(this),
 			this.checkButtonStatus.bind(this)
+		)
+		this.feedbacks = new FragmentFeedbacks(
+			registry,
+			controlId,
+			this.commitChange.bind(this),
+			this.triggerRedraw.bind(this),
+			false
 		)
 	}
 
@@ -156,323 +157,14 @@ export default class ButtonControlBase extends ControlBase {
 	}
 
 	/**
-	 * Inform the instance of a removed feedback
-	 * @param {object} fedeback the feedback being removed
-	 * @access private
-	 */
-	#cleanupFeedback(feedback) {
-		// Inform relevant module
-		const instance = this.instance.moduleHost.getChild(feedback.instance_id)
-		if (instance) {
-			instance.feedbackDelete(feedback).catch((e) => {
-				this.logger.silly(`feedback_delete to connection failed: ${e.message}`)
-			})
-		}
-
-		// Remove from cached feedback values
-		delete this.#cachedFeedbackValues[feedback.id]
-	}
-
-	/**
 	 * Prepare this control for deletion
 	 * @access public
 	 */
 	destroy() {
-		// TODO - some of this feels like it could be done in bulk
-
-		// Inform modules of feedback cleanup
-		for (const feedback of this.feedbacks) {
-			this.#cleanupFeedback(feedback)
-		}
-
 		this.actions.destroy()
+		this.feedbacks.destroy()
 
 		super.destroy()
-	}
-
-	/**
-	 * Add a feedback to this control
-	 * @param {object} feedbackItem the item to add
-	 * @returns {boolean} success
-	 * @access public
-	 */
-	feedbackAdd(feedbackItem) {
-		this.feedbacks.push(feedbackItem)
-
-		// Inform relevant module
-		this.#feedbackSubscribe(feedbackItem)
-
-		this.commitChange()
-
-		return true
-	}
-
-	/**
-	 * Duplicate an feedback on this control
-	 * @param {string} id
-	 * @returns {boolean} success
-	 * @access public
-	 */
-	feedbackDuplicate(id) {
-		const index = this.feedbacks.findIndex((fb) => fb.id === id)
-		if (index !== -1) {
-			const feedbackItem = cloneDeep(this.feedbacks[index])
-			feedbackItem.id = nanoid()
-
-			this.feedbacks.splice(index + 1, 0, feedbackItem)
-
-			this.#feedbackSubscribe(feedbackItem)
-
-			this.commitChange(false)
-
-			return true
-		}
-
-		return false
-	}
-
-	feedbackEnabled(id, enabled) {
-		for (const feedback of this.feedbacks) {
-			if (feedback && feedback.id === id) {
-				if (!feedback.options) feedback.options = {}
-
-				feedback.disabled = !enabled
-
-				// Remove from cached feedback values
-				delete this.#cachedFeedbackValues[id]
-
-				// Inform relevant module
-				if (!feedback.disabled) {
-					this.#feedbackSubscribe(feedback)
-				} else {
-					this.#cleanupFeedback(feedback)
-				}
-
-				this.commitChange()
-
-				return true
-			}
-		}
-
-		return false
-	}
-
-	/**
-	 * Learn the options for a feedback, by asking the instance for the current values
-	 * @param {string} id the id of the feedback
-	 * @returns {boolean} success
-	 * @access public
-	 */
-	async feedbackLearn(id) {
-		const feedback = this.feedbacks.find((fb) => fb.id === id)
-		if (feedback) {
-			const instance = this.instance.moduleHost.getChild(feedback.instance_id)
-			if (instance) {
-				const newOptions = await instance.feedbackLearnValues(feedback)
-				if (newOptions) {
-					const newFeedback = {
-						...feedback,
-						options: newOptions,
-					}
-
-					// It may not still exist, so do a replace through the usual flow
-					return this.feedbackReplace(newFeedback)
-				}
-			}
-		}
-
-		return false
-	}
-
-	/**
-	 * Remove a feedback from this control
-	 * @param {string} id the id of the feedback
-	 * @returns {boolean} success
-	 * @access public
-	 */
-	feedbackRemove(id) {
-		const index = this.feedbacks.findIndex((fb) => fb.id === id)
-		if (index !== -1) {
-			const feedback = this.feedbacks[index]
-			this.feedbacks.splice(index, 1)
-
-			this.#cleanupFeedback(feedback)
-
-			this.commitChange()
-
-			return true
-		} else {
-			return false
-		}
-	}
-
-	/**
-	 * Reorder a feedback in the list
-	 * @param {number} oldIndex the index of the feedback to move
-	 * @param {number} newIndex the target index of the feedback
-	 * @returns {boolean} success
-	 * @access public
-	 */
-	feedbackReorder(oldIndex, newIndex) {
-		oldIndex = clamp(oldIndex, 0, this.feedbacks.length)
-		newIndex = clamp(newIndex, 0, this.feedbacks.length)
-		this.feedbacks.splice(newIndex, 0, ...this.feedbacks.splice(oldIndex, 1))
-
-		this.commitChange()
-	}
-
-	/**
-	 * Replace a feedback with an updated version
-	 * @param {object} newProps
-	 * @access public
-	 */
-	feedbackReplace(newProps) {
-		for (const feedback of this.feedbacks) {
-			// Replace the new feedback in place
-			if (feedback.id === newProps.id) {
-				feedback.type = newProps.feedbackId
-				feedback.options = newProps.options
-
-				delete feedback.upgradeIndex
-
-				// Preserve existing value if it is set
-				feedback.style = feedback.style || newProps.style
-
-				this.#feedbackSubscribe(feedback)
-
-				this.commitChange(true)
-
-				return true
-			}
-		}
-
-		return false
-	}
-
-	/**
-	 * Update an option for a feedback
-	 * @param {string} id the id of the feedback
-	 * @param {string} key the key/name of the property
-	 * @param {any} value the new value
-	 * @returns {boolean} success
-	 * @access public
-	 */
-	feedbackSetOptions(id, key, value) {
-		for (const feedback of this.feedbacks) {
-			if (feedback && feedback.id === id) {
-				if (!feedback.options) feedback.options = {}
-
-				feedback.options[key] = value
-
-				// Remove from cached feedback values
-				delete this.#cachedFeedbackValues[id]
-
-				// Inform relevant module
-				this.#feedbackSubscribe(feedback)
-
-				this.commitChange()
-
-				return true
-			}
-		}
-
-		return false
-	}
-
-	/**
-	 * Update the selected style properties for a boolean feedback
-	 * @param {string} id the id of the feedback
-	 * @param {string[]} selected the properties to be selected
-	 * @returns {boolean} success
-	 * @access public
-	 */
-	feedbackSetStyleSelection(id, selected) {
-		for (const feedback of this.feedbacks) {
-			if (feedback && feedback.id === id) {
-				const definition = this.instance.definitions.getFeedbackDefinition(feedback.instance_id, feedback.type)
-				if (!definition || definition.type !== 'boolean') return false
-
-				const defaultStyle = definition.style || {}
-				const oldStyle = feedback.style || {}
-				const newStyle = {}
-
-				for (const key of selected) {
-					if (key in oldStyle) {
-						// preserve existing value
-						newStyle[key] = oldStyle[key]
-					} else {
-						// copy bank value, as a default
-						newStyle[key] = defaultStyle[key] !== undefined ? defaultStyle[key] : this.style[key]
-
-						// png needs to be set to something harmless
-						if (key === 'png64' && !newStyle[key]) {
-							newStyle[key] = null
-						}
-					}
-				}
-				feedback.style = newStyle
-
-				this.commitChange()
-
-				return true
-			}
-		}
-
-		return false
-	}
-
-	/**
-	 * Update an style property for a boolean feedback
-	 * @param {string} id the id of the feedback
-	 * @param {string} key the key/name of the property
-	 * @param {any} value the new value
-	 * @returns {boolean} success
-	 * @access public
-	 */
-	feedbackSetStyleValue(id, key, value) {
-		if (key === 'png64') {
-			if (!value.match(/data:.*?image\/png/)) {
-				return false
-			}
-
-			value = value.replace(/^.*base64,/, '')
-		}
-
-		for (const feedback of this.feedbacks) {
-			if (feedback && feedback.id === id) {
-				const definition = this.instance.definitions.getFeedbackDefinition(feedback.instance_id, feedback.type)
-				if (!definition || definition.type !== 'boolean') return false
-
-				if (!feedback.style) feedback.style = {}
-				feedback.style[key] = value
-
-				this.commitChange()
-
-				return true
-			}
-		}
-
-		return false
-	}
-
-	/**
-	 * Inform the instance of an updated feedback
-	 * @param {object} feedback the feedback which changed
-	 * @access private
-	 */
-	#feedbackSubscribe(feedback) {
-		if (!feedback.disabled) {
-			if (feedback.instance_id === 'internal') {
-				this.internalModule.feedbackUpdate(feedback, this.controlId)
-			} else {
-				const instance = this.instance.moduleHost.getChild(feedback.instance_id)
-				if (instance) {
-					instance.feedbackUpdate(feedback, this.controlId).catch((e) => {
-						this.logger.silly(`feedback_update to connection failed: ${e.message} ${e.stack}`)
-					})
-				}
-			}
-		}
 	}
 
 	/**
@@ -481,20 +173,7 @@ export default class ButtonControlBase extends ControlBase {
 	 * @access public
 	 */
 	forgetInstance(instanceId) {
-		let changedFeedbacks = false
-
-		// Cleanup any feedbacks
-		const newFeedbacks = []
-		for (const feedback of this.feedbacks) {
-			if (feedback.instance_id === instanceId) {
-				this.#cleanupFeedback(feedback)
-				changedFeedbacks = true
-			} else {
-				newFeedbacks.push(feedback)
-			}
-		}
-		this.feedbacks = newFeedbacks
-
+		const changedFeedbacks = this.feedbacks.forgetInstance(instanceId)
 		const changedActions = this.actions.forgetInstance(instanceId)
 
 		if (changedActions || changedFeedbacks) {
@@ -540,30 +219,10 @@ export default class ButtonControlBase extends ControlBase {
 	 * @access public
 	 */
 	getUnparsedStyle() {
-		let style = { ...this.style }
-
-		// Iterate through feedback-overrides
-		for (const feedback of this.feedbacks) {
-			if (feedback.disabled) continue
-
-			const definition = this.instance.definitions.getFeedbackDefinition(feedback.instance_id, feedback.type)
-			const rawValue = this.#cachedFeedbackValues[feedback.id]
-			if (definition && rawValue !== undefined) {
-				if (definition.type === 'boolean' && rawValue == true) {
-					style = {
-						...style,
-						...feedback?.style,
-					}
-				} else if (definition.type === 'advanced' && typeof rawValue === 'object') {
-					style = {
-						...style,
-						...rawValue,
-					}
-				}
-			}
+		return {
+			...this.style,
+			...this.feedbacks.getUnparsedStyleOverrides(),
 		}
-
-		return style
 	}
 
 	/**
@@ -605,24 +264,11 @@ export default class ButtonControlBase extends ControlBase {
 	postProcessImport() {
 		const ps = []
 
-		// Make sure all the ids are unique
-		for (const feedback of this.feedbacks) {
-			feedback.id = nanoid()
-
-			if (feedback.instance_id === 'internal') {
-				this.internalModule.feedbackUpdate(feedback, this.controlId)
-			} else {
-				const instance = this.instance.moduleHost.getChild(feedback.instance_id)
-				if (instance) {
-					ps.push(instance.feedbackUpdate(feedback, this.controlId))
-				}
-			}
-		}
-
+		ps.push(this.feedbacks.postProcessImport())
 		ps.push(this.actions.postProcessImport())
 
 		Promise.all(ps).catch((e) => {
-			this.logger.silly(`posProcessImport for ${this.controlId} failed: ${e.message}`)
+			this.logger.silly(`postProcessImport for ${this.controlId} failed: ${e.message}`)
 		})
 
 		this.commitChange()
@@ -724,9 +370,7 @@ export default class ButtonControlBase extends ControlBase {
 
 			if ('show_topbar' in diff) {
 				// Some feedbacks will need to redraw
-				for (const feedback of this.feedbacks) {
-					this.#feedbackSubscribe(feedback)
-				}
+				this.feedbacks.resubscribeAllFeedbacks()
 			}
 
 			this.commitChange()
@@ -738,50 +382,16 @@ export default class ButtonControlBase extends ControlBase {
 	}
 
 	/**
-	 * Update the feedbacks on the button with new values
-	 * @param {string} instanceId The instance the feedbacks are for
-	 * @param {object} newValues The new fedeback values
-	 */
-	updateFeedbackValues(instanceId, newValues) {
-		let changed = false
-
-		for (const feedback of this.feedbacks) {
-			if (feedback.instance_id === instanceId) {
-				if (feedback.id in newValues) {
-					// Feedback is present in new values (might be set to undefined)
-					const value = newValues[feedback.id]
-					if (!isEqual(value, this.#cachedFeedbackValues[feedback.id])) {
-						// Found the feedback, exactly where it said it would be
-						// Mark the bank as changed, and store the new value
-						this.#cachedFeedbackValues[feedback.id] = value
-						changed = true
-					}
-				}
-			}
-		}
-
-		if (changed) {
-			this.triggerRedraw()
-		}
-	}
-
-	/**
 	 * Prune all actions/feedbacks referencing unknown instances
 	 * Doesn't do any cleanup, as it is assumed that the instance has not been running
 	 * @param {Set<string>} knownInstanceIds
 	 * @access public
 	 */
 	verifyInstanceIds(knownInstanceIds) {
-		let changed = false
-
-		// Clean out feedbacks
-		const feedbackLength = this.feedbacks.length
-		this.feedbacks = this.feedbacks.filter((feedback) => !!feedback && knownInstanceIds.has(feedback.instance_id))
-		changed = changed || this.feedbacks.length !== feedbackLength
-
+		const changedFeedbacks = this.feedbacks.verifyInstanceIds(knownInstanceIds)
 		const changedActions = this.actions.verifyInstanceIds(knownInstanceIds)
 
-		if (changed || changedActions) {
+		if (changedFeedbacks || changedActions) {
 			this.commitChange()
 		}
 	}
