+++ lib/Controls/ControlTypes/Button/Press.js
@@ -37,7 +37,7 @@ export default class ControlButtonPress extends ButtonControlBase {
 
 		this.style = cloneDeep(ButtonControlBase.DefaultStyle)
 		this.options = cloneDeep(ButtonControlBase.DefaultOptions)
-		this.feedbacks = []
+		this.feedbacks.feedbacks = []
 		this.actions.action_sets = {
 			down: [],
 			up: [],
@@ -54,7 +54,7 @@ export default class ControlButtonPress extends ButtonControlBase {
 
 			this.style = storage.style || this.style
 			this.options = storage.options || this.options
-			this.feedbacks = storage.feedbacks || this.feedbacks
+			this.feedbacks.feedbacks = storage.feedbacks || this.feedbacks.feedbacks
 			this.actions.action_sets = storage.action_sets || this.actions.action_sets
 
 			if (isImport) this.postProcessImport()
@@ -96,7 +96,7 @@ export default class ControlButtonPress extends ButtonControlBase {
 			type: this.type,
 			style: this.style,
 			options: this.options,
-			feedbacks: this.feedbacks,
+			feedbacks: this.feedbacks.feedbacks,
 			action_sets: this.actions.action_sets,
 		}
 		return clone ? cloneDeep(obj) : obj
