+++ lib/Controls/ControlTypes/Button/Step.js
@@ -46,7 +46,7 @@ export default class ControlButtonStep extends ButtonControlBase {
 			...cloneDeep(ButtonControlBase.DefaultOptions),
 			stepAutoProgress: true,
 		}
-		this.feedbacks = []
+		this.feedbacks.feedbacks = []
 		this.actions.action_sets = {
 			0: [],
 		}
@@ -63,7 +63,7 @@ export default class ControlButtonStep extends ButtonControlBase {
 
 			this.style = storage.style || this.style
 			this.options = storage.options || this.options
-			this.feedbacks = storage.feedbacks || this.feedbacks
+			this.feedbacks.feedbacks = storage.feedbacks || this.feedbacks.feedbacks
 			this.actions.action_sets = storage.action_sets || this.actions.action_sets
 
 			this.#current_step_id = Object.keys(this.actions.action_sets)[0]
@@ -303,7 +303,7 @@ export default class ControlButtonStep extends ButtonControlBase {
 			type: this.type,
 			style: this.style,
 			options: this.options,
-			feedbacks: this.feedbacks,
+			feedbacks: this.feedbacks.feedbacks,
 			action_sets: this.actions.action_sets,
 		}
 		return clone ? cloneDeep(obj) : obj
