+++ lib/Controls/ControlTypes/Triggers/Events/Timer.js
@@ -39,7 +39,7 @@ export default class TriggersEventTimer {
 			time = `${Math.floor(event.options.seconds / 60)} minutes`
 		}
 
-		return `Runs every <strong>${time}</strong>.`
+		return `Every <strong>${time}</strong>.`
 	}
 
 	getTimeOfDayDescription(event) {
