+++ lib/Controls/ControlTypes/Triggers/Trigger.js
@@ -327,7 +327,7 @@ export default class ControlTrigger extends ControlBase {
 		ps.push(this.actions.postProcessImport())
 
 		Promise.all(ps).catch((e) => {
-			this.logger.silly(`posProcessImport for ${this.controlId} failed: ${e.message}`)
+			this.logger.silly(`postProcessImport for ${this.controlId} failed: ${e.message}`)
 		})
 
 		this.commitChange()
