+++ lib/Controls/Controller.js
@@ -245,10 +245,14 @@ class ControlsController extends CoreBase {
 			const control = this.getControl(controlId)
 			if (!control) return false
 
-			if (typeof control.feedbackAdd === 'function') {
-				const feedbackItem = this.instance.definitions.createFeedbackItem(instanceId, feedbackId)
+			if (control.feedbacks && typeof control.feedbacks.feedbackAdd === 'function') {
+				const feedbackItem = this.instance.definitions.createFeedbackItem(
+					instanceId,
+					feedbackId,
+					control.feedbacks.isBooleanOnly
+				)
 				if (feedbackItem) {
-					return control.feedbackAdd(feedbackItem)
+					return control.feedbacks.feedbackAdd(feedbackItem)
 				} else {
 					return false
 				}
@@ -261,8 +265,8 @@ class ControlsController extends CoreBase {
 			const control = this.getControl(controlId)
 			if (!control) return false
 
-			if (typeof control.feedbackLearn === 'function') {
-				return control.feedbackLearn(id)
+			if (control.feedbacks && typeof control.feedbacks.feedbackLearn === 'function') {
+				return control.feedbacks.feedbackLearn(id)
 			} else {
 				throw new Error(`Control "${controlId}" does not support feedbacks`)
 			}
@@ -272,8 +276,8 @@ class ControlsController extends CoreBase {
 			const control = this.getControl(controlId)
 			if (!control) return false
 
-			if (typeof control.feedbackEnabled === 'function') {
-				return control.feedbackEnabled(id, enabled)
+			if (control.feedbacks && typeof control.feedbacks.feedbackEnabled === 'function') {
+				return control.feedbacks.feedbackEnabled(id, enabled)
 			} else {
 				throw new Error(`Control "${controlId}" does not support feedbacks`)
 			}
@@ -283,8 +287,8 @@ class ControlsController extends CoreBase {
 			const control = this.getControl(controlId)
 			if (!control) return false
 
-			if (typeof control.feedbackRemove === 'function') {
-				return control.feedbackRemove(id)
+			if (control.feedbacks && typeof control.feedbacks.feedbackRemove === 'function') {
+				return control.feedbacks.feedbackRemove(id)
 			} else {
 				throw new Error(`Control "${controlId}" does not support feedbacks`)
 			}
@@ -294,8 +298,8 @@ class ControlsController extends CoreBase {
 			const control = this.getControl(controlId)
 			if (!control) return false
 
-			if (typeof control.feedbackDuplicate === 'function') {
-				return control.feedbackDuplicate(id)
+			if (control.feedbacks && typeof control.feedbacks.feedbackDuplicate === 'function') {
+				return control.feedbacks.feedbackDuplicate(id)
 			} else {
 				throw new Error(`Control "${controlId}" does not support feedbacks`)
 			}
@@ -305,8 +309,8 @@ class ControlsController extends CoreBase {
 			const control = this.getControl(controlId)
 			if (!control) return false
 
-			if (typeof control.feedbackSetOptions === 'function') {
-				return control.feedbackSetOptions(id, key, value)
+			if (control.feedbacks && typeof control.feedbacks.feedbackSetOptions === 'function') {
+				return control.feedbacks.feedbackSetOptions(id, key, value)
 			} else {
 				throw new Error(`Control "${controlId}" does not support feedbacks`)
 			}
@@ -316,8 +320,8 @@ class ControlsController extends CoreBase {
 			const control = this.getControl(controlId)
 			if (!control) return false
 
-			if (typeof control.feedbackReorder === 'function') {
-				return control.feedbackReorder(oldIndex, newIndex)
+			if (control.feedbacks && typeof control.feedbacks.feedbackReorder === 'function') {
+				return control.feedbacks.feedbackReorder(oldIndex, newIndex)
 			} else {
 				throw new Error(`Control "${controlId}" does not support feedbacks`)
 			}
@@ -326,8 +330,8 @@ class ControlsController extends CoreBase {
 			const control = this.getControl(controlId)
 			if (!control) return false
 
-			if (typeof control.feedbackSetStyleSelection === 'function') {
-				return control.feedbackSetStyleSelection(id, selected)
+			if (control.feedbacks && typeof control.feedbacks.feedbackSetStyleSelection === 'function') {
+				return control.feedbacks.feedbackSetStyleSelection(id, selected)
 			} else {
 				throw new Error(`Control "${controlId}" does not support feedbacks`)
 			}
@@ -336,8 +340,8 @@ class ControlsController extends CoreBase {
 			const control = this.getControl(controlId)
 			if (!control) return false
 
-			if (typeof control.feedbackSetStyleValue === 'function') {
-				return control.feedbackSetStyleValue(id, key, value)
+			if (control.feedbacks && typeof control.feedbacks.feedbackSetStyleValue === 'function') {
+				return control.feedbacks.feedbackSetStyleValue(id, key, value)
 			} else {
 				throw new Error(`Control "${controlId}" does not support feedbacks`)
 			}
@@ -913,8 +917,8 @@ class ControlsController extends CoreBase {
 		// Pass values to controls
 		for (const [controlId, newValues] of Object.entries(values)) {
 			const control = this.getControl(controlId)
-			if (control && typeof control.updateFeedbackValues === 'function') {
-				control.updateFeedbackValues(instanceId, newValues)
+			if (control && control.feedbacks && typeof control.feedbacks.updateFeedbackValues === 'function') {
+				control.feedbacks.updateFeedbackValues(instanceId, newValues)
 			}
 		}
 	}
