+++ lib/Controls/Fragments/FragmentActions.js
@@ -449,7 +449,7 @@ export default class FragmentActions extends CoreBase {
 		}
 
 		await Promise.all(ps).catch((e) => {
-			this.logger.silly(`posProcessImport for ${this.controlId} failed: ${e.message}`)
+			this.logger.silly(`postProcessImport for ${this.controlId} failed: ${e.message}`)
 		})
 	}
