+++ lib/Data/ImportExport.js
@@ -123,7 +123,7 @@ class DataImportExport extends CoreBase {
 				const instance_ids = new Set()
 				for (const control of Object.values(exp.controls)) {
 					// Future: this should be done inside of the control classes, as they could use a non-standard structure
-					for (const feedback of control.feedbacks || []) {
+					for (const feedback of control.feedbacks.feedbacks || []) {
 						instance_ids.add(feedback.instance_id)
 					}
 
@@ -528,8 +528,8 @@ class DataImportExport extends CoreBase {
 			}
 			control.feedbacks = newFeedbacks
 		}
-		if (control.actions && control.actions.action_sets) {
-			for (const [setId, action_set] of Object.entries(control.actions.action_sets)) {
+		if (control.action_sets) {
+			for (const [setId, action_set] of Object.entries(control.action_sets)) {
 				const newActions = []
 				for (const action of action_set) {
 					const instanceInfo = instanceIdMap[action?.instance]
@@ -539,7 +539,7 @@ class DataImportExport extends CoreBase {
 						newActions.push(action)
 					}
 				}
-				control.actions.action_sets[setId] = newActions
+				control.action_sets[setId] = newActions
 			}
 		}
