+++ lib/Instance/Definitions.js
@@ -174,11 +174,14 @@ class InstanceDefinitions extends CoreBase {
 	 * Create a feedback item without saving fpr the UI
 	 * @param {string} instanceId - the id of the instance
 	 * @param {string} feedbackId - the id of the feedback
+	 * @param {boolean} booleanOnly - whether the feedback must be boolean
 	 * @access public
 	 */
-	createFeedbackItem(instanceId, feedbackId) {
+	createFeedbackItem(instanceId, feedbackId, booleanOnly) {
 		const definition = this.getFeedbackDefinition(instanceId, feedbackId)
 		if (definition) {
+			if (booleanOnly && definition.type !== 'boolean') return null
+
 			const feedback = {
 				id: nanoid(),
 				type: feedbackId,
@@ -194,7 +197,7 @@ class InstanceDefinitions extends CoreBase {
 				}
 			}
 
-			if (definition.type === 'boolean' && definition.style) {
+			if (!booleanOnly && definition.type === 'boolean' && definition.style) {
 				feedback.style = cloneDeep(definition.style)
 			}
