+++ lib/Instance/Wrapper.js
@@ -113,9 +113,9 @@ class SocketEventsHandler {
 		// Find all the feedbacks on banks
 		const allControls = this.registry.controls.getAllControls()
 		for (const [controlId, control] of Object.entries(allControls)) {
-			if (control.feedbacks && control.feedbacks.length > 0) {
+			if (control.feedbacks && control.feedbacks.feedbacks && control.feedbacks.feedbacks.length > 0) {
 				const imageSize = control.getBitmapSize()
-				for (const feedback of control.feedbacks) {
+				for (const feedback of control.feedbacks.feedbacks) {
 					const parsed = ParseControlId(controlId)
 					if (feedback.instance_id === this.connectionId && parsed?.type === 'bank') {
 						allFeedbacks[feedback.id] = {
@@ -585,7 +585,7 @@ class SocketEventsHandler {
 		for (const feedback of Object.values(msg.updatedFeedbacks)) {
 			if (feedback) {
 				const control = this.registry.controls.getControl(feedback.controlId)
-				const found = control?.feedbackReplace?.(feedback) ?? false
+				const found = control?.feedbacks?.feedbackReplace?.(feedback) ?? false
 				if (!found) {
 					this.logger.silly(`Failed to replace upgraded feedback: ${feedback.id} ${feedback.controlId}`)
 				}
