+++ lib/Internal/Controller.js
@@ -57,8 +57,8 @@ export default class InternalController extends CoreBase {
 		const allControls = this.registry.controls.getAllControls()
 		for (const [controlId, control] of Object.entries(allControls)) {
 			// Discover feedbacks to process
-			if (control.feedbacks) {
-				for (const feedback of control.feedbacks) {
+			if (control.feedbacks && control.feedbacks.feedbacks) {
+				for (const feedback of control.feedbacks.feedbacks) {
 					if (feedback.instance_id === 'internal') {
 						this.feedbackUpdate(feedback, controlId)
 					}
